+++ formula/src/main/java/com/instacart/formula/FormulaContext.kt
@@ -21,7 +21,7 @@ abstract class FormulaContext<State> internal constructor(
      * It uses [transition] type as key.
      */
     fun <Event> onEvent(
-        transition: Transition.Factory.(Event) -> Transition<State>,
+        transition: Transition<State, Event>,
     ): Listener<Event> {
         return onEvent(Unit, transition)
     }
@@ -33,7 +33,7 @@ abstract class FormulaContext<State> internal constructor(
      */
     fun <Event> onEvent(
         key: Any,
-        transition: Transition.Factory.(Event) -> Transition<State>,
+        transition: Transition<State, Event>,
     ): Listener<Event> {
         val type = transition::class
         val joinedKey: Any = if (key == Unit) {
@@ -52,7 +52,7 @@ abstract class FormulaContext<State> internal constructor(
      *
      * It uses [transition] type as key.
      */
-    fun callback(transition: Transition.Factory.(Unit) -> Transition<State>): () -> Unit {
+    fun callback(transition: Transition<State, Unit>): () -> Unit {
         val listener = onEvent(transition)
         return UnitListener(listener)
     }
@@ -64,7 +64,7 @@ abstract class FormulaContext<State> internal constructor(
      */
     fun callback(
         key: Any,
-        transition: Transition.Factory.(Unit) -> Transition<State>
+        transition: Transition<State, Unit>,
     ): () -> Unit {
         val listener = onEvent(key, transition)
         return UnitListener(listener)
@@ -76,7 +76,7 @@ abstract class FormulaContext<State> internal constructor(
      * It uses [transition] type as key.
      */
     fun <Event> eventCallback(
-        transition: Transition.Factory.(Event) -> Transition<State>
+        transition: Transition<State, Event>,
     ): Listener<Event> {
         return onEvent(transition)
     }
@@ -88,7 +88,7 @@ abstract class FormulaContext<State> internal constructor(
      */
     fun <Event> eventCallback(
         key: Any,
-        transition: Transition.Factory.(Event) -> Transition<State>,
+        transition: Transition<State, Event>,
     ): Listener<Event> {
         return onEvent(key, transition)
     }
@@ -129,5 +129,4 @@ abstract class FormulaContext<State> internal constructor(
         listeners.endScope()
         return value
     }
-
 }