+++ formula/src/main/java/com/instacart/formula/Transition.kt
@@ -1,79 +1,66 @@
 package com.instacart.formula
 
 /**
- * Defines an intent to transition by emitting a new [State] and optional [Effects].
+ * Transition is a function that is called when an [Event] happens and produces a [Result] which
+ * indicates what [Formula] should do in response to this event. It can contain a new [state][State]
+ * object which will trigger [Formula.evaluate] to be called and/or an [effects][Effects] function
+ * which will be executed by the [FormulaRuntime]. If there was a state change, effects function
+ * will be executed after [Formula.evaluate] is called.
  */
-sealed class Transition<out State> {
-    companion object {
+fun interface Transition<State, in Event> {
+
+    /**
+     * Result is an object returned by [Transition.toResult] which indicates what
+     * [Formula] should do in response to an event. It can contain a new [state][Stateful]
+     * and/or an [effects][Effects] function which will be executed by the [FormulaRuntime].
+     * Usually, you should use [TransitionContext] to construct the [Result] object.
+     */
+    sealed class Result<out State> {
+
         /**
-         * A convenience method to define transitions.
+         * Stateful result returned by the transition indicates that we want to apply a
+         * new state to our [Formula]. It can also return optional [effects] function
+         * which will be executed by [FormulaRuntime] after the state change.
          *
-         * ```
-         * fun nameChanged(state: FormState, newName: String) = Transition.create {
-         *   transition(state.copy(name = newName))
-         * }
-         * ```
+         * @param state New state which will be applied to our [Formula].
+         * @param effects Optional function which will be executed after the state change. This
+         * function is the place to call listeners, log events, trigger network requests or
+         * database writes, and etc.
          */
-        inline fun <State> create(init: Factory.() -> Transition<State>): Transition<State> {
-            return init(Factory)
-        }
-    }
+        data class Stateful<State>(val state: State, override val effects: Effects? = null) : Result<State>()
 
-    /**
-     * Stateful transition.
-     *
-     * @param state New state
-     * @param effects Optional effects such as calling listeners, logging, db writes,
-     * network requests, etc.
-     */
-    data class Stateful<State>(val state: State, override val effects: Effects? = null) : Transition<State>()
-
-    /**
-     * Only effects are emitted as part of this transition
-     *
-     * @param effects Effects such as calling listeners, logging, db writes, network requests, etc.
-     */
-    data class OnlyEffects(override val effects: Effects) : Transition<Nothing>()
-
-    /**
-     * Nothing happens in this transition.
-     */
-    object None : Transition<Nothing>() {
-        override val effects: Effects? = null
-    }
-
-
-    /**
-     * Factory uses as a receiver parameter to provide transition constructor dsl.
-     */
-    object Factory {
         /**
-         * A transition that does nothing.
+         * Only effects result returned by the transition indicates that we don't need to update
+         * [Formula] state and only want [effects] function to be executed by [FormulaRuntime].
+         *
+         * @param effects A function which will be executed by [FormulaRuntime]. This function
+         * is the place to call listeners, log events, trigger network requests or database
+         * writes, and etc.
          */
-        fun none(): Transition<Nothing> {
-            return None
+        data class OnlyEffects(override val effects: Effects) : Result<Nothing>()
+
+        /**
+         * None result returned by the transition indicates that [Formula] doesn't need to
+         * do anything in response to an event.
+         */
+        object None : Result<Nothing>() {
+            override val effects: Effects? = null
         }
 
         /**
-         * Creates a transition to a new [State] and executes [invokeEffects] function
-         * after the state change.
+         * Effects is an optional function returned by [toResult] which will be executed by
+         * [FormulaRuntime]. This function is the place to call listeners, log events, trigger
+         * network requests or database writes, and etc.
+         *
+         * @see Effects
          */
-        fun <State> transition(
-            state: State,
-            invokeEffects: (() -> Unit)? = null
-        ): Stateful<State> {
-            return Stateful(state, invokeEffects)
-        }
-
-        /**
-         * Creates a transition that only executes [invokeEffects].
-         */
-        fun transition(
-            invokeEffects: () -> Unit
-        ): OnlyEffects {
-            return OnlyEffects(invokeEffects)
-        }
+        abstract val effects: Effects?
     }
 
-    abstract val effects: Effects?
+    /**
+     * Called when an [Event] happens and returns a [Result] object which can indicate a state
+     * change and/or some executable effects. Use [TransitionContext.none] if nothing should happen
+     * as part of this event.
+     */
+    fun TransitionContext<State>.toResult(event: Event): Result<State>
 }