+++ formula/src/main/java/com/instacart/formula/internal/TransitionListener.kt
@@ -3,5 +3,5 @@ package com.instacart.formula.internal
 import com.instacart.formula.Transition
 
 internal fun interface TransitionListener {
-    fun onTransition(transition: Transition<*>, isValid: Boolean)
+    fun onTransitionResult(result: Transition.Result<*>, isValid: Boolean)
 }