+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Add `FormulaContext.onEvent` to replace `FormulaContext.callback` and `FormulaContext.eventCallback`
 - Add Kotlin coroutine support! New `formula-coroutines` module to provide `toFlow()` extension function and `FlowStream`.
 - Add `Listener` type to consolidate event handling types `() -> Unit` and `(Event) -> Unit` 
-- Rename and move `FormulaContext.UpdateBuilder` into `StreamBuilder`. 
+- **Breaking**: Rename and move `FormulaContext.UpdateBuilder` into `StreamBuilder`. 
+- **Breaking**: Restructuring `Transition` type into `Transition<State, Event>` and `Transition.Result<State>`
 
 ## [0.7.0] - June 30, 2021
 - **Breaking**: Remove `events(observable) { }` utility function.